First pass over generic map
